fix: Capturer: Fix the formatting issue
This is the problem with all of this conditional compliation - cargo
ignores these files entirely and I miss things like this. I have to do
all kinds of `#[cfg(target_os)]` shenanigans to work around it for
development, and then try to undo them all before committing. So error
prone.