feat: Config: Move config file location
This is a breaking change, so warrants a major version bump.

Let's just put the config file in the users home directory.
We can also put everything hanging off of there by default as well,
but if you ever want to put the shots & videos somewhere else you'll be
left with `~/ompd/config.json` and nothing else in there.
That's just silly.
So one config file in your home directory that says where to find
everything else.