fix: Capturer: acquire screen every time
At any given point while running the screen containing (0, 0) may
change. A great example would be when plugging a laptop into an external
display which becomes the primary. Conversely, if we start with the
external display and then it's removed. It's all the same to us.

WWhen that happens we don't automagically catpure The Right Thing - and
so our captures end up being little 1x1 pixel nonsense - but not errors!

So, acquire the screen holding (0, 0) every time. This is generally
fine, though on Linux it's not that efficient.

This actually happens in the [display-info crate](https://github.com/nashaofu/display-info)